added branch longer. switch int -> long in SalesManager.java

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {1, 2, 5, 3};
+        long[] sales = {1, 2, 5, 3};
         SalesManager manager1 = new SalesManager(sales);
         System.out.println("Максимальная сумма продажи " + manager1.max());
